fix bug in all.css
The all.css file used in template homepage overwritted the all.css file used in other pages. But the file does not have same requirements on css files. 

In development environment, this was not seeable. But in production environment, the all.css file used in most of the page was not complete, preventing a correct display.